Rename files according to React/JSX styleguide
